Generacion con bounding funcionando
