--- H3.py
@@ -1,5 +1,0 @@
-import h3
-# import googlemaps
-lat, lng = 37.769377, -122.388903
-resolution = 9
-h3.latlng_to_cell(lat, lng, resolution)
