+++ requirements.txt
@@ -1,3 +1,5 @@
 folium~=0.18.0
 geopandas~=1.0.1
 python-dotenv~=1.0.1
+shapely~=2.0.6
+branca~=0.8.0
